txnsync: add message acknowledge  (#2219)
This PR adds the transaction sync message acknowledgement : whenever a peer receive a message and add its entire content to its transaction pool, the message sequence number would be sent back. The caller would add the previously sent transactions IDs to a long-term cache, indicating that there is no need to resend these messages.

The advantage in doing so, is that we don't need to sent back a bloom filter once we receive a message. This reduces the amount of verbosity dramatically.

+++ txnsync/exchange.go
@@ -63,6 +63,6 @@ type timingParams struct {
 
 	RefTxnBlockMsgSeq   uint64   `codec:"s"`
 	ResponseElapsedTime uint64   `codec:"r"`
-	AcceptedMsgSeq      []uint32 `codec:"a,allocbound=maxAcceptedMsgSeq"`
+	AcceptedMsgSeq      []uint64 `codec:"a,allocbound=maxAcceptedMsgSeq"`
 	NextMsgMinDelay     uint64   `codec:"m"`
 }
